-Changing float to double when using Trackball::distance.

+++ trunk/OsgTools/Render/Viewer.h
@@ -639,11 +639,11 @@ protected:
   
   virtual osg::Vec3d         getCenter();
   virtual void               setCenter( const osg::Vec3d& );
-  virtual float              getDistance();
-  virtual void               setDistance ( float );
+  virtual double             getDistance();
+  virtual void               setDistance ( double );
   virtual osg::Quat          getRotation();
   virtual void               setRotation( const osg::Quat& );
-  virtual void               setTrackball ( const osg::Vec3d&, float distance, const osg::Quat&, bool makeTrackball, bool setViewerToo );
+  virtual void               setTrackball ( const osg::Vec3d&, double distance, const osg::Quat&, bool makeTrackball, bool setViewerToo );
 
   /////////////////////////////////////////////////////////////////////////////
   //
